refactor(run): shuffle around files in cycle/run
Simply moves code around, no other significant changes. Tests were split into three files, one for
each API. Implementation was split between internals.ts and index.ts
